+++ packages/application/src/ingest/index.ts
@@ -12,3 +12,4 @@ export * from "./process-ingest-job";
 export * from "./process-parse-job";
 export * from "./delete-document";
 export * from "./reparse-document";
+export * from "./prune-document-versions";